requirements.txt
streamlit==1.22.0
ultralytics==8.0.20
opencv-python-headless==4.5.5.64
Pillow>=8.0.0
torch>=1.7.0
numpy>=1.21.0
